week4

+++ Assignment1/tuples.h
@@ -8,16 +8,16 @@ using namespace std;
 // https://www.geeksforgeeks.org/shuffle-a-given-array-using-fisher-yates-shuffle-algorithm/
 
 // A utility function to swap to integers 
-void swap (tuple<int8_t,int8_t> *a, tuple<int8_t,int8_t> *b) 
+void swap (tuple<int64_t,int64_t> *a, tuple<int64_t,int64_t> *b) 
 { 
-    tuple<int8_t,int8_t> temp = *a; 
+    tuple<int64_t,int64_t> temp = *a; 
     *a = *b; 
     *b = temp; 
 } 
 
 // A function to generate a random 
 // permutation of arr[] 
-void randomize (tuple<int8_t,int8_t> arr[], int n) 
+void randomize (tuple<int64_t,int64_t> arr[], int n) 
 { 
     // Use a different seed value so that 
     // we don't get same result each time
@@ -39,8 +39,8 @@ void randomize (tuple<int8_t,int8_t> arr[], int n)
 } 
 
 // num_Tuples = 2^24
-tuple<int8_t,int8_t>* makeInput(size_t numTuples) {
-    tuple<int8_t, int8_t>* input = new tuple<int8_t, int8_t>[numTuples];
+tuple<int64_t,int64_t>* makeInput(size_t numTuples) {
+    tuple<int64_t, int64_t>* input = new tuple<int64_t, int64_t>[numTuples];
     
     for (int i = 0; i <numTuples; i++) {
         input[i] = std::make_tuple(i,0);
@@ -52,6 +52,6 @@ tuple<int8_t,int8_t>* makeInput(size_t numTuples) {
 }
 
 // modular hash function
-int hashFunction(int8_t key, int numPartitions ) {
+int hashFunction(int64_t key, int numPartitions ) {
     return key % numPartitions;
 }
